+++ src/axom/quest/tests/quest_gwn_methods.cpp
@@ -155,7 +155,7 @@ TEST(quest_gwn_methods, gwn_moment_data_triangle)
 
 //------------------------------------------------------------------------------
 template <typename ExecSpace>
-void check_mfem_mesh_linearization()
+void check_mfem_mesh_evaluation()
 {
   using NURBSCurve2D = axom::primal::NURBSCurve<double, 2>;
   const std::string fileName = pjoin(AXOM_DATA_DIR, "contours", "svg", "mfem_logo_simp.mesh");
@@ -168,7 +168,7 @@ void check_mfem_mesh_linearization()
   const int ret = mfem_reader.read(curves);
   if(ret != 0)
   {
-    SLIC_ERROR(axom::fmt::format("Failed to read STEP file '{}'", fileName));
+    SLIC_ERROR(axom::fmt::format("Failed to read mesh file '{}'", fileName));
   }
 
   // Get a linearization of the shape
@@ -187,12 +187,12 @@ void check_mfem_mesh_linearization()
   const std::vector<int> resolution {10, 10};
   const int query_order = 1;
 
-  // Generate three query grids and fields
-  axom::Array<mfem::DataCollection> dc(0, 3);
-  std::string names[] = {"direct", "polyline", "polyline_fast"};
-  for(int i = 0; i < 3; ++i)
+  // Generate query grids and fields
+  constexpr int num_queries = 6;
+  axom::Array<mfem::DataCollection> dc(0, num_queries);
+  for(int i = 0; i < num_queries; ++i)
   {
-    dc.emplace_back(axom::fmt::format("gwn_{}", names[i]));
+    dc.emplace_back(axom::fmt::format("gwn_method_{}", i));
     axom::quest::generate_gwn_query_mesh(dc[i],
                                          shape_bbox,
                                          std::vector<double> {},
@@ -203,51 +203,61 @@ void check_mfem_mesh_linearization()
 
   // Create tolerance object
   axom::primal::WindingTolerances tol;
-  constexpr bool useDirectPolyline = true;
+  constexpr bool useDirectEvaluation = true;
+  constexpr bool useMemoization = true;
 
-  //// Run three different kinds of GWN query ////
+  //// Run six different kinds of GWN query ////
   // We expect all three fields to return the same values in this case because
   //  of the specific arrangement of query points and linearization.
   // In general, discretizing the shape can result in different GWN values
   //  for query points near to individual curves.
 
-  // Direct
-  SLIC_INFO("Testing Direct Evaluation");
-  axom::quest::DirectGWN2D<ExecSpace> gwn_direct {};
-  gwn_direct.preprocess(curves);
-  gwn_direct.query(dc[0], tol);
+  SLIC_INFO("Testing Curve Evaluation");
+  axom::quest::NURBSCurveGWNQuery<ExecSpace> gwn_curves {};
+  gwn_curves.preprocess(curves, useDirectEvaluation, !useMemoization);
+  gwn_curves.query(dc[0], tol);
 
-  // Linearized
-  SLIC_INFO("Testing Direct Evaluation of Triangulation");
-  axom::quest::PolylineGWN2D<ExecSpace, 0> gwn_polyline {};
-  gwn_polyline.preprocess(&poly_mesh, useDirectPolyline);
-  gwn_polyline.query(dc[1], tol);
+  axom::quest::NURBSCurveGWNQuery<ExecSpace> gwn_curves_memoized {};
+  gwn_curves_memoized.preprocess(curves, useDirectEvaluation, useMemoization);
+  gwn_curves_memoized.query(dc[1], tol);
 
-  // Linearized, fast approximation
-  SLIC_INFO("Testing Fast-Approximate Evaluation of Triangulation");
-  axom::quest::PolylineGWN2D<ExecSpace, 0> gwn_polyline_fast {};
-  gwn_polyline_fast.preprocess(&poly_mesh, !useDirectPolyline);
-  gwn_polyline_fast.query(dc[2], tol);
+  axom::quest::NURBSCurveGWNQuery<ExecSpace, 0> gwn_curves_fast {};
+  gwn_curves_fast.preprocess(curves, !useDirectEvaluation, !useMemoization);
+  gwn_curves_fast.query(dc[2], tol);
 
-  // Compare the in-out values between the three fields
+  axom::quest::NURBSCurveGWNQuery<ExecSpace, 0> gwn_curves_fast_memoized {};
+  gwn_curves_fast_memoized.preprocess(curves, !useDirectEvaluation, useMemoization);
+  gwn_curves_fast_memoized.query(dc[3], tol);
+
+  SLIC_INFO("Testing Linearization Evaluation");
+  axom::quest::PolylineGWNQuery<ExecSpace> gwn_polyline {};
+  gwn_polyline.preprocess(&poly_mesh, useDirectEvaluation);
+  gwn_polyline.query(dc[4], tol);
+
+  axom::quest::PolylineGWNQuery<ExecSpace, 0> gwn_polyline_fast {};
+  gwn_polyline_fast.preprocess(&poly_mesh, !useDirectEvaluation);
+  gwn_polyline_fast.query(dc[5], tol);
+
+  // Compare the in-out values between all fields
   const auto *query_mesh = dc[0].GetMesh();
   const auto num_query_points = query_mesh->GetNodalFESpace()->GetNDofs();
 
   auto &inout_direct = *dc[0].GetField("inout");
-  auto &inout_polyline = *dc[1].GetField("inout");
-  auto &inout_polyline_fast = *dc[2].GetField("inout");
-
-  for(int i = 0; i < num_query_points; ++i)
+  for(int N = 1; N < num_queries; ++N)
   {
-    EXPECT_EQ(inout_direct[i], inout_polyline[i]);
-    EXPECT_EQ(inout_direct[i], inout_polyline_fast[i]);
+    auto &inout_other = *dc[N].GetField("inout");
+
+    for(int i = 0; i < num_query_points; ++i)
+    {
+      EXPECT_EQ(inout_direct[i], inout_other[i]);
+    }
   }
 }
 
 #ifdef AXOM_USE_OPENCASCADE
 //------------------------------------------------------------------------------
 template <typename ExecSpace>
-void check_step_file_triangulation()
+void check_step_file_evaluation()
 {
   const std::string fileName = pjoin(AXOM_DATA_DIR, "quest", "step", "nut.step");
 
@@ -275,11 +285,11 @@ void check_step_file_triangulation()
   const int query_order = 1;
 
   // Generate three query grids and fields
-  axom::Array<mfem::DataCollection> dc(0, 3);
-  std::string names[] = {"direct", "tri", "tri_fast"};
-  for(int i = 0; i < 3; ++i)
+  constexpr int num_queries = 4;
+  axom::Array<mfem::DataCollection> dc(0, num_queries);
+  for(int i = 0; i < num_queries; ++i)
   {
-    dc.emplace_back(axom::fmt::format("gwn_{}", names[i]));
+    dc.emplace_back(axom::fmt::format("gwn_method_{}", i));
     axom::quest::generate_gwn_query_mesh(dc[i],
                                          shape_bbox,
                                          std::vector<double> {},
@@ -290,73 +300,74 @@ void check_step_file_triangulation()
 
   // Create tolerance object
   axom::primal::WindingTolerances tol;
-  constexpr bool useDirectTriangle = true;
+  constexpr bool useDirectEvaluation = true;
+  constexpr bool useMemoization = true;
 
-  //// Run three different kinds of GWN query ////
+  //// Run four different kinds of GWN query ////
   // We expect all three fields to return the same values in this case because
   //  of the specific arrangement of query points and triangulation.
   // In general, triangulating the shape can result in different GWN values
   //  for query points near to individual surfaces.
 
-  // Direct
-  SLIC_INFO("Testing Direct Evaluation");
-  axom::quest::DirectGWN3D<ExecSpace> gwn_direct {};
-  gwn_direct.preprocess(patches);
-  gwn_direct.query(dc[0], tol);
+  // Direct patch evaluation is prohibitively slow without memoization.
+  //  Keep interface only for symmetry with NURBSCurve methods
 
-  // Triangulated
-  SLIC_INFO("Testing Direct Evaluation of Polyline");
-  axom::quest::TriangleGWN3D<ExecSpace, 0> gwn_tri {};
-  gwn_tri.preprocess(&tri_mesh, useDirectTriangle);
-  gwn_tri.query(dc[1], tol);
+  SLIC_INFO("Testing Patch Evaluation");
+  //axom::quest::NURBSPatchGWNQuery<ExecSpace> gwn_patches {};
+  //gwn_patches.preprocess(patches, useDirectEvaluation, !useMemoization);
+  //gwn_patches.query(dc[0], tol);
 
-  // Triangulated, fast approximation
-  SLIC_INFO("Testing Fast-Approximate Evaluation of Polyline");
-  axom::quest::TriangleGWN3D<ExecSpace, 0> gwn_tri_fast {};
-  gwn_tri_fast.preprocess(&tri_mesh, !useDirectTriangle);
-  gwn_tri_fast.query(dc[2], tol);
+  axom::quest::NURBSPatchGWNQuery<ExecSpace> gwn_patches_memoized {};
+  gwn_patches_memoized.preprocess(patches, useDirectEvaluation, useMemoization);
+  gwn_patches_memoized.query(dc[0], tol);
 
-  // Compare the in-out values between the three fields
+  //axom::quest::NURBSPatchGWNQuery<ExecSpace, 0> gwn_patches_fast {};
+  //gwn_patches_fast.preprocess(patches, !useDirectEvaluation, !useMemoization);
+  //gwn_patches_fast.query(dc[2], tol);
+
+  axom::quest::NURBSPatchGWNQuery<ExecSpace, 0> gwn_patches_fast_memoized {};
+  gwn_patches_fast_memoized.preprocess(patches, !useDirectEvaluation, useMemoization);
+  gwn_patches_fast_memoized.query(dc[1], tol);
+
+  SLIC_INFO("Testing Triangulation Evaluation");
+  axom::quest::TriangleGWNQuery<ExecSpace> gwn_triangles {};
+  gwn_triangles.preprocess(&tri_mesh, useDirectEvaluation);
+  gwn_triangles.query(dc[2], tol);
+
+  axom::quest::TriangleGWNQuery<ExecSpace, 0> gwn_triangles_fast {};
+  gwn_triangles_fast.preprocess(&tri_mesh, !useDirectEvaluation);
+  gwn_triangles_fast.query(dc[3], tol);
+
+  // Compare the in-out values between all fields
   const auto *query_mesh = dc[0].GetMesh();
   const auto num_query_points = query_mesh->GetNodalFESpace()->GetNDofs();
 
   auto &inout_direct = *dc[0].GetField("inout");
-  auto &inout_tri = *dc[1].GetField("inout");
-  auto &inout_tri_fast = *dc[2].GetField("inout");
-
-  for(int i = 0; i < num_query_points; ++i)
+  for(int N = 1; N < num_queries; ++N)
   {
-    EXPECT_EQ(inout_direct[i], inout_tri[i]);
-    EXPECT_EQ(inout_direct[i], inout_tri_fast[i]);
+    auto &inout_other = *dc[N].GetField("inout");
+
+    for(int i = 0; i < num_query_points; ++i)
+    {
+      EXPECT_EQ(inout_direct[i], inout_other[i]);
+    }
   }
 }
 #endif
 
 //------------------------------------------------------------------------------
-TEST(quest_gwn_methods, mfem_mesh_linearization)
-{
-  check_mfem_mesh_linearization<axom::SEQ_EXEC>();
-}
+TEST(quest_gwn_methods, mfem_mesh_evaluation) { check_mfem_mesh_evaluation<axom::SEQ_EXEC>(); }
 
 #if defined(AXOM_USE_OPENMP) && defined(AXOM_USE_RAJA)
-TEST(quest_gwn_methods, mfem_mesh_linearization_omp)
-{
-  check_mfem_mesh_linearization<axom::OMP_EXEC>();
-}
+TEST(quest_gwn_methods, mfem_mesh_evaluation_omp) { check_mfem_mesh_evaluation<axom::OMP_EXEC>(); }
 #endif
 
 #ifdef AXOM_USE_OPENCASCADE
-TEST(quest_gwn_methods, step_file_triangulation)
-{
-  check_step_file_triangulation<axom::SEQ_EXEC>();
-}
+TEST(quest_gwn_methods, step_file_evaluation) { check_step_file_evaluation<axom::SEQ_EXEC>(); }
 #endif
 
 #if defined(AXOM_USE_OPENCASCADE) && defined(AXOM_USE_OPENMP) && defined(AXOM_USE_RAJA)
-TEST(quest_gwn_methods, step_file_triangulation_omp)
-{
-  check_step_file_triangulation<axom::OMP_EXEC>();
-}
+TEST(quest_gwn_methods, step_file_evaluation_omp) { check_step_file_evaluation<axom::OMP_EXEC>(); }
 #endif
 
 //------------------------------------------------------------------------------